Refine terminator matching logic and extract some more to match_algorithms
