+++ src/sqlfluff/core/parser/grammar/base.py
@@ -5,7 +5,6 @@ from typing import (
     TYPE_CHECKING,
     Any,
     Callable,
-    Iterable,
     List,
     Optional,
     Sequence,
@@ -18,6 +17,7 @@ from uuid import UUID, uuid4
 
 from sqlfluff.core.parser.context import ParseContext
 from sqlfluff.core.parser.helpers import trim_non_code_segments
+from sqlfluff.core.parser.match_algorithms import prune_options
 from sqlfluff.core.parser.match_logging import parse_match_logging
 from sqlfluff.core.parser.match_result import MatchResult
 from sqlfluff.core.parser.match_wrapper import match_wrapper
@@ -224,78 +224,6 @@ class BaseGrammar(Matchable):
         """Does this matcher support a lowercase hash matching route?"""
         return None
 
-    @staticmethod
-    def _first_non_whitespace(
-        segments: Iterable["BaseSegment"],
-    ) -> Optional[Tuple[str, Set[str]]]:
-        """Return the upper first non-whitespace segment in the iterable."""
-        for segment in segments:
-            if segment.first_non_whitespace_segment_raw_upper:
-                return (
-                    segment.first_non_whitespace_segment_raw_upper,
-                    segment.class_types,
-                )
-        return None
-
-    @classmethod
-    def _prune_options(
-        cls,
-        options: List[MatchableType],
-        segments: Tuple[BaseSegment, ...],
-        parse_context: ParseContext,
-    ) -> List[MatchableType]:
-        """Use the simple matchers to prune which options to match on.
-
-        Works in the context of a grammar making choices between options
-        such as AnyOf or the content of Delimited.
-        """
-        available_options = []
-        prune_buff = []
-
-        # Find the first code element to match against.
-        first_segment = cls._first_non_whitespace(segments)
-        # If we don't have an appropriate option to match against,
-        # then we should just return immediately. Nothing will match.
-        if not first_segment:
-            return options
-        first_raw, first_types = first_segment
-
-        for opt in options:
-            simple = opt.simple(parse_context=parse_context)
-            if simple is None:
-                # This element is not simple, we have to do a
-                # full match with it...
-                available_options.append(opt)
-                continue
-
-            # Otherwise we have a simple option, so let's use
-            # it for pruning.
-            simple_raws, simple_types = simple
-            matched = False
-
-            # We want to know if the first meaningful element of the str_buff
-            # matches the option, based on either simple _raw_ matching or
-            # simple _type_ matching.
-
-            # Match Raws
-            if simple_raws and first_raw in simple_raws:
-                # If we get here, it's matched the FIRST element of the string buffer.
-                available_options.append(opt)
-                matched = True
-
-            # Match Types
-            if simple_types and not matched and first_types.intersection(simple_types):
-                # If we get here, it's matched the FIRST element of the string buffer.
-                available_options.append(opt)
-                matched = True
-
-            if not matched:
-                # Ditch this option, the simple match has failed
-                prune_buff.append(opt)
-                continue
-
-        return available_options
-
     @classmethod
     def _longest_trimmed_match(
         cls,
@@ -344,7 +272,7 @@ class BaseGrammar(Matchable):
             return MatchResult.from_unmatched(segments), None
 
         # Prune available options, based on their simple representation for efficiency.
-        available_options = cls._prune_options(
+        available_options = prune_options(
             matchers, segments, parse_context=parse_context
         )
